reorganised the src to app and views
