+++ app/Helpers/csrf_functions.php
