+++ app/Services/impl/AccountServiceImpl.php
@@ -2,12 +2,13 @@
 
 namespace App\Services\impl;
 
+use App\Core\View;
 use App\Services\AccountService;
 
 class AccountServiceImpl implements AccountService
 {
     public function profileView()
     {
-        require_once 'views/account.php';
+        View::render('account');
     }
 }
