+++ composer.json
@@ -1,20 +1,20 @@
 {
     "autoload": {
         "psr-4": {
-            "App\\Controllers\\": "src/Controllers/",
-            "App\\Dtos\\": "src/Dtos/",
-            "App\\Models\\": "src/Models/",
-            "App\\Repositories\\": "src/Repositories/",
-            "App\\Services\\": "src/Services/",
-            "App\\" : "src/"
+            "App\\Controllers\\": "app/Controllers/",
+            "App\\Dtos\\": "app/Dtos/",
+            "App\\Models\\": "app/Models/",
+            "App\\Repositories\\": "app/Repositories/",
+            "App\\Services\\": "app/Services/",
+            "App\\" : "app/"
         },
         "files": [
-            "src/Helpers/csrf_functions.php",
-            "src/Helpers/flash_messages.php",
-            "src/Helpers/old_field_functions.php",
-            "src/Helpers/route_functions.php",
-            "src/Helpers/user_functions.php",
-            "src/Helpers/validation_functions.php"
+            "app/Helpers/csrf_functions.php",
+            "app/Helpers/flash_messages.php",
+            "app/Helpers/old_field_functions.php",
+            "app/Helpers/route_functions.php",
+            "app/Helpers/user_functions.php",
+            "app/Helpers/validation_functions.php"
         ]
     },
     "require": {
